Also moved enum Trap out of vm
