+++ lib/vm/src/trap/mod.rs
@@ -6,8 +6,8 @@
 mod traphandlers;
 
 pub use traphandlers::{
-    catch_traps, on_host_stack, raise_lib_trap, raise_user_trap, wasmer_call_trampoline, Trap,
+    catch_traps, on_host_stack, raise_lib_trap, raise_user_trap, wasmer_call_trampoline,
     TrapHandler, TrapHandlerFn,
 };
 pub use traphandlers::{init_traps, resume_panic};
-pub use wasmer_types::TrapCode;
+pub use wasmer_types::{Trap, TrapCode};
